Models/pyproject.toml
[project]
name = "models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "groq>=0.37.1",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "transformers>=4.57.6",
]
